Refactor application structure and enhance RabbitMQ integration: Updated environment configurations in .env.example, CLAUDE.md, and docker-compose.yml to support new RabbitMQ setup. Introduced separate Dockerfiles for publisher and consumer services, along with entrypoint scripts for initialization. Removed outdated QStash configurations and refactored database schema for transaction management. Added deployment scripts and improved documentation for local development and deployment processes.

+++ migrations/001_create_transactions_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS transactions (
     id SERIAL PRIMARY KEY,
     transaction_id VARCHAR(255) UNIQUE NOT NULL,
-    value DECIMAL(15,2) NOT NULL,
+    value BIGINT NOT NULL,
     type VARCHAR(10) NOT NULL CHECK (type IN ('in', 'out')),
     external_company_id VARCHAR(255) NOT NULL,
     description TEXT,
